Refactor into packages capybara and command
