--- Data/taskStorage.txt
@@ -1,7 +1,0 @@
-T | 0 | sleep
-D | 0 | sleep | 2025-06-04 22:00
-D | 0 | sleep | 2025-06-04
-E | 1 | nap | 2025-06-04 18:00 | 2025-06-05
-T | 0 | sleep
-D | 0 | sleep | 2025-06-04
-D | 0 | nap | 2025-06-04 18:00
