+++ src/main/java/capybara/CapyException.java
@@ -1,0 +1,15 @@
+package capybara;
+
+public class CapyException extends Exception {
+    public CapyException(String message) {
+        super(message);
+    }
+}
+
+
+
+
+
+
+
+
