+++ src/main/java/capybara/Deadline.java
@@ -1,6 +1,6 @@
+package capybara;
+
 import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 
 public class Deadline extends Task {
     private final LocalDateTime by;
