+++ src/main/java/capybara/EmptyTaskNumberException.java
@@ -1,0 +1,7 @@
+package capybara;
+
+public class EmptyTaskNumberException extends CapyException {
+    EmptyTaskNumberException(String kind) {
+        super("OOPS!!! capybara.Capybara requires a task number to " + kind + " :-(");
+    }
+}
