+++ src/main/java/capybara/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package capybara;
+
+public class UnknownCommandException extends CapyException {
+    UnknownCommandException(String cmd) {
+        super("OOPS!!! I'm sorry, but I don't know what that means :-(");
+    }
+}
