+++ src/main/java/capybara/command/AddDeadlineCommand.java
@@ -1,11 +1,14 @@
-/*
 package capybara.command;
 
 import capybara.CapyException;
+import capybara.Deadline;
 import capybara.Storage;
+import capybara.Task;
 import capybara.TaskList;
 import capybara.Ui;
-*/
+
+
+import capybara.*;
 
 import java.time.LocalDateTime;
