+++ src/main/java/capybara/command/ByeCommand.java
@@ -1,15 +1,17 @@
-/*
 package capybara.command;
 
-import capybara.CapyException;
 import capybara.Storage;
 import capybara.TaskList;
 import capybara.Ui;
-*/
 
-public class ListCommand extends Command {
+public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showList(tasks.formatAll());
+        ui.showGoodbye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
     }
 }
