+++ src/main/java/capybara/command/Command.java
@@ -1,11 +1,10 @@
-/*
 package capybara.command;
 
 import capybara.CapyException;
 import capybara.Storage;
 import capybara.TaskList;
 import capybara.Ui;
-*/
+
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage)
