+++ src/main/java/capybara/command/ListCommand.java
@@ -1,20 +1,12 @@
-/*
 package capybara.command;
 
-import capybara.CapyException;
 import capybara.Storage;
 import capybara.TaskList;
 import capybara.Ui;
-*/
 
-public class ByeCommand extends Command {
+public class ListCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showGoodbye();
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
+        ui.showList(tasks.formatAll());
     }
 }
