+++ text-ui-test/EXPECTED.TXT
@@ -4,7 +4,7 @@ ____________________________________________________________
   (-┬- -┬-)       
   (_______)       
 
- Hello! I'm Capybara
+ Hello! I'm capybara.Capybara
  What can I do for you... Zzzzz
 ____________________________________________________________
 ____________________________________________________________
